Finished the rewrite of the D3D12 layer
