+++ src/Titan/App.cs
@@ -1,16 +1,9 @@
 using Titan.Application;
-using Titan.Configurations;
-using Titan.Core.IO.Platform;
 using Titan.Core.Logging;
 using Titan.Core.Memory;
-using Titan.Core.Memory.Platform;
 using Titan.Core.Threading;
-using Titan.Core.Threading.Platform;
 using Titan.Events;
-using Titan.IO.FileSystem;
 using Titan.Modules;
-using Titan.Resources;
-using Titan.Systems;
 
 namespace Titan;
