+++ src/Titan/Application/TitanApp.cs
@@ -1,4 +1,3 @@
-using Microsoft.Win32;
 using System.Diagnostics;
 using Titan.Configurations;
 using Titan.Core;
