+++ src/Titan/Rendering/D3D12/Adapters/D3D12Adapter.cs
@@ -1,39 +1,47 @@
+using Titan.Configurations;
 using Titan.Core;
 using Titan.Core.Logging;
-using Titan.Core.Memory;
 using Titan.Platform.Win32;
 using Titan.Platform.Win32.DXGI;
+using Titan.Resources;
+using Titan.Systems;
 using static Titan.Platform.Win32.Win32Common;
 
 namespace Titan.Rendering.D3D12.Adapters;
-internal sealed unsafe class D3D12Adapter : IService
-{
-    private TitanArray<DXGIAdapter> _adapters;
-    private IMemoryManager? _memoryManager;
 
-    private uint _primaryAdapterIndex = 0;
-    public ref readonly DXGIAdapter PrimaryAdapter => ref _adapters[_primaryAdapterIndex];
-    public bool Init(IMemoryManager memoryManager, AdapterConfig? config, bool debug)
+[UnmanagedResource]
+internal unsafe partial struct D3D12Adapter
+{
+    public const int MaxAdapters = 10;
+    public uint PrimaryAdapterIndex;
+    public uint AdapterCount;
+    public Inline8<AdapterInfo> Adapters;
+    public readonly ref readonly AdapterInfo PrimaryAdapter => ref Adapters[PrimaryAdapterIndex];
+
+
+    [System(SystemStage.Init)]
+    public static void Init(D3D12Adapter* adapter, IConfigurationManager configurationManager)
     {
-        var flags = debug ? DXGI_CREATE_FACTORY_FLAGS.DXGI_CREATE_FACTORY_DEBUG : 0;
+        var config = configurationManager.GetConfigOrDefault<RenderingConfig>();
+
+        var flags = config.Debug ? DXGI_CREATE_FACTORY_FLAGS.DXGI_CREATE_FACTORY_DEBUG : 0;
         Logger.Trace<D3D12Adapter>($"Creating {nameof(IDXGIFactory7)}. Flags = {flags}");
         using ComPtr<IDXGIFactory7> factory = default;
         var hr = DXGICommon.CreateDXGIFactory2(flags, factory.UUID, (void**)factory.GetAddressOf());
         if (FAILED(hr))
         {
             Logger.Error<D3D12Adapter>($"Failed to create a {nameof(IDXGIFactory7)}. HRESULT = {hr}");
-            return false;
+            //TODO(Jens): Implement fatal error
+            return;
         }
 
-        const int maxAdapters = 10;
-        Span<DXGIAdapter> adapters = stackalloc DXGIAdapter[maxAdapters];
-        adapters.Clear();
+        //NOTE(Jens): We use the first device we find if there's no stored config. We enumerate them by GPU preference so we expect to get the best one first.
+        adapter->PrimaryAdapterIndex = 0;
 
-        uint index;
-        for (index = 0u; index < maxAdapters; ++index)
+        for (var index = 0u; index < MaxAdapters; ++index)
         {
-            IDXGIAdapter3* adapter = null;
-            hr = factory.Get()->EnumAdapterByGpuPreference(index, DXGI_GPU_PREFERENCE.DXGI_GPU_PREFERENCE_HIGH_PERFORMANCE, IDXGIAdapter3.Guid, (void**)&adapter);
+            IDXGIAdapter3* dxgiAdapter;
+            hr = factory.Get()->EnumAdapterByGpuPreference(index, DXGI_GPU_PREFERENCE.DXGI_GPU_PREFERENCE_HIGH_PERFORMANCE, IDXGIAdapter3.Guid, (void**)&dxgiAdapter);
             if (hr == DXGI_ERROR.DXGI_ERROR_NOT_FOUND)
             {
                 break;
@@ -45,67 +53,58 @@ internal sealed unsafe class D3D12Adapter : IService
             }
 
             DXGI_ADAPTER_DESC1 desc;
-            hr = adapter->GetDesc1(&desc);
+            hr = dxgiAdapter->GetDesc1(&desc);
             if (FAILED(hr))
             {
                 Logger.Error<D3D12Adapter>($"Failed to get the adapter desc. Index = {index} HRESULT = {hr}");
+                dxgiAdapter->Release();
                 break;
             }
 
-            adapters[(int)index] = new(adapter, desc);
-            Logger.Trace<D3D12Adapter>($"Found adapter {adapters[(int)index].DebugString}");
+            ref var info = ref adapter->Adapters[adapter->AdapterCount++];
+            info.Adapter = dxgiAdapter;
+            info.SetName(desc.DescriptionString());
+            info.DeviceId = desc.DeviceId;
+            info.VendorId = desc.VendorId;
+            info.IsHardware = (desc.Flags & DXGI_ADAPTER_FLAG.DXGI_ADAPTER_FLAG_SOFTWARE) == 0;
+
+            Logger.Trace<D3D12Adapter>($"Found adapter {info.DebugString}");
         }
 
-        if (index == 0)
+        if (config.Adapter != null)
         {
-            Logger.Error<D3D12Adapter>("No adapters found.");
-            return false;
-        }
-
-        if (!memoryManager.TryAllocArray(out _adapters, index))
-        {
-            Logger.Error<D3D12Adapter>($"Failed to allocate array. Count = {index} Size = {sizeof(DXGIAdapter) * index}");
-            return false;
-        }
-
-        adapters[..(int)index]
-            .CopyTo(_adapters.AsSpan());
-
-        _primaryAdapterIndex = GetPrimaryAdapterIndex(config);
-
-        _memoryManager = memoryManager;
-
-        return true;
-    }
-
-    private uint GetPrimaryAdapterIndex(AdapterConfig? config)
-    {
-        if (config == null)
-        {
-            Logger.Trace<D3D12Adapter>("No adapter configuration");
-            return 0;
-        }
-
-        for (var i = 0u; i < _adapters.Length; ++i)
-        {
-            ref readonly var desc = ref _adapters[i].Desc;
-            if (desc.DeviceId == config.DeviceId && desc.VendorId == config.VendorId)
+            var deviceId = config.Adapter.DeviceId;
+            var vendorId = config.Adapter.VendorId;
+            Logger.Info<D3D12Adapter>($"Using configured adapter. DeviceId = {deviceId} VendorId = {vendorId}");
+            var found = false;
+            for (var i = 0u; i < adapter->AdapterCount; ++i)
             {
-                Logger.Trace<D3D12Adapter>($"Found a matching Adapter. Name = {_adapters[i].Name} DeviceId = {config.DeviceId} VendorId = {config.VendorId}");
-                return i;
+                ref readonly var info = ref adapter->Adapters[i];
+                if (info.DeviceId == deviceId && info.VendorId == vendorId)
+                {
+                    adapter->PrimaryAdapterIndex = i;
+                    found = true;
+                    break;
+                }
+            }
+
+            if (!found)
+            {
+                Logger.Warning<D3D12Adapter>($"The adapter in the configuration could not be found, using default device. Configured DeviceId = {deviceId} Configured VendorId = {vendorId}");
             }
         }
 
-        Logger.Trace<D3D12Adapter>($"No matching adapter found. DeviceId = {config.DeviceId} VendorId = {config.VendorId}");
-        return 0;
+        Logger.Trace<D3D12Adapter>($"Found {adapter->AdapterCount} adapters.");
     }
 
-    public void Shutdown()
+    [System(SystemStage.Shutdown)]
+    public static void Shutdown(D3D12Adapter* adapter)
     {
-        foreach (ref var adapter in _adapters.AsSpan())
+        for (var i = 0; i < adapter->AdapterCount; ++i)
         {
-            adapter.Dispose();
+            adapter->Adapters[i].Adapter.Dispose();
         }
-        _memoryManager?.FreeArray(ref _adapters);
+
+        *adapter = default;
     }
 }
