+++ src/Titan/Rendering/D3D12/Memory/D3D12Allocator.cs
@@ -1,196 +1,132 @@
 using System.Diagnostics;
-using System.Runtime.CompilerServices;
+using Titan.Configurations;
+using Titan.Core;
 using Titan.Core.Logging;
 using Titan.Core.Memory;
-using Titan.Platform.Win32;
 using Titan.Platform.Win32.D3D12;
-using Titan.Platform.Win32.DXGI;
-using Titan.Rendering.D3D12.Utils;
+using Titan.Resources;
+using Titan.Systems;
 
 namespace Titan.Rendering.D3D12.Memory;
 
-internal unsafe struct TempBufferDescriptor
+[UnmanagedResource]
+internal unsafe partial struct D3D12Allocator
 {
-    public ComPtr<ID3D12Resource> Resource;
-    public void* CPUAddress;
-    public ulong GPUAddress;
-    public volatile uint Offset;
-    public DescriptorHandle SRV;
-}
+    public const uint BufferCount = GlobalConfiguration.MaxRenderFrames;
 
-internal sealed unsafe class D3D12Allocator : IService
-{
-    private const uint ConstantBufferAlignment = 256;
-    private const uint BufferCount = GlobalConfiguration.MaxRenderFrames;
-    private DescriptorHeaps _heaps;
-    private TempBuffersDescriptors _tempBuffers;
-    private D3D12Device? _device;
-    private IMemoryManager? _memoryManager;
+    private TitanArray<int> _sharedFreeList;
+    private Inline4<DescriptorHeap> _heaps;
+
     private int _frameIndex;
-    private uint _tempBufferSize;
-    public bool Init(IMemoryManager memoryManager, D3D12Device device, in GPUMemoryConfig config)
+
+    [System(SystemStage.Init)]
+    public static void Init(D3D12Allocator* allocator, in D3D12Device device, IMemoryManager memoryManager, IConfigurationManager configurationManager)
     {
-        for (var i = 0; i < (int)DescriptorHeapType.Count; ++i)
+        var config = configurationManager.GetConfigOrDefault<D3D12Config>();
+        var memoryConfig = config.MemoryConfig;
+
+        if (!memoryManager.TryAllocArray(out allocator->_sharedFreeList, memoryConfig.TotalCount))
         {
-            var type = (DescriptorHeapType)i;
-            var tempCount = type == DescriptorHeapType.ShaderResourceView ? config.TempShaderResourceViewCount : 0;
-            var shaderVisible = type == DescriptorHeapType.ShaderResourceView;
-            var count = config.GetDescriptorCount(type);
-            if (!_heaps[i].Init(memoryManager, device, type, count, tempCount, shaderVisible))
-            {
-                Logger.Error<D3D12Allocator>($"Failed to initialize the {nameof(DescriptorHeap)}. Type = {type} Count = {count} ShaderVisibile = {shaderVisible} TempCount = {tempCount}");
-                return false;
-            }
+            Logger.Error<D3D12Allocator>($"Failed to allocate memory for the shared free list. Count = {memoryConfig.TotalCount} Size = {memoryConfig.TotalCount * sizeof(uint)}");
+            return;
         }
 
-        // Init the temp buffers
-        var tempBufferSize = MemoryUtils.AlignToUpper(config.TempBufferSize, ConstantBufferAlignment);
-        for (var i = 0; i < BufferCount; ++i)
+        var offset = 0u;
+        for (DescriptorHeapType type = 0; type < DescriptorHeapType.Count; ++type)
         {
-            var resource = device.CreateBuffer(tempBufferSize, true, D3D12_RESOURCE_STATES.D3D12_RESOURCE_STATE_VERTEX_AND_CONSTANT_BUFFER);
-            if (resource == null)
+            var d3d12Type = type switch
             {
-                Logger.Error<D3D12Allocator>($"Failed to create the temporary buffer. Size = {tempBufferSize} bytes Index = {i}");
-                return false;
-            }
-            D3D12Helpers.SetName(resource, $"TempConstantBuffer_{i}");
-            var desc = new TempBufferDescriptor
-            {
-                Resource = resource,
-                GPUAddress = resource->GetGPUVirtualAddress(),
-                SRV = Allocate(DescriptorHeapType.ShaderResourceView)
+                DescriptorHeapType.ShaderResourceView => D3D12_DESCRIPTOR_HEAP_TYPE.D3D12_DESCRIPTOR_HEAP_TYPE_CBV_SRV_UAV,
+                DescriptorHeapType.DepthStencilView => D3D12_DESCRIPTOR_HEAP_TYPE.D3D12_DESCRIPTOR_HEAP_TYPE_DSV,
+                DescriptorHeapType.RenderTargetView => D3D12_DESCRIPTOR_HEAP_TYPE.D3D12_DESCRIPTOR_HEAP_TYPE_RTV,
+                DescriptorHeapType.UnorderedAccessView => D3D12_DESCRIPTOR_HEAP_TYPE.D3D12_DESCRIPTOR_HEAP_TYPE_CBV_SRV_UAV,
+                _ => throw new InvalidOperationException("Not supported")
             };
-            var hr = resource->Map(0, null, &desc.CPUAddress);
-            if (Win32Common.FAILED(hr))
+
+            ref var heap = ref allocator->_heaps[(int)type];
+
+            heap.NumberOfDescriptors = memoryConfig.GetDescriptorCount(type);
+            heap.FreeList = allocator->_sharedFreeList.Slice(offset, heap.NumberOfDescriptors);
+            heap.IncrementSize = device.GetDescriptorHandleIncrementSize(d3d12Type);
+            heap.ShaderVisible = type is DescriptorHeapType.ShaderResourceView;
+            //NOTE(Jens): We only have temporary SRV descriptors.
+            heap.NumberOfTempDescriptors = type is DescriptorHeapType.ShaderResourceView ? memoryConfig.TempShaderResourceViewCount : 0;
+
+
+            var totalCount = BufferCount * heap.NumberOfTempDescriptors + heap.NumberOfDescriptors;
+            ID3D12DescriptorHeap* descriptorHeap = heap.Heap = device.CreateDescriptorHeap(d3d12Type, totalCount, heap.ShaderVisible);
+            if (descriptorHeap == null)
             {
-                Logger.Error<D3D12Allocator>($"Failed to map the temporary buffer. Index = {i} HRESULT = {hr}");
-                return false;
+                Logger.Error<D3D12Allocator>($"Failed to create the {nameof(ID3D12DescriptorHeap)}. Type = {type} D3D12Type = {d3d12Type} Count = {totalCount} ShaderVisible = {heap.ShaderVisible}");
+                // fatal
             }
-            _tempBuffers[i] = desc;
-        }
 
-        _memoryManager = memoryManager;
-        _device = device;
-        _tempBufferSize = tempBufferSize;
-
-
-        return true;
-    }
-
-    [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    public DescriptorHandle Allocate(DescriptorHeapType type)
-    {
-        Debug.Assert(type != DescriptorHeapType.Count);
-        return _heaps[type].Allocate();
-    }
-
-    public void Free(in DescriptorHandle handle)
-    {
-        Debug.Assert(handle.Type != DescriptorHeapType.Count);
-        _heaps[handle.Type].Free(handle);
-    }
-
-
-    [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    public StructuredBuffer<T> AllocateTempStructuredBuffer<T>(int count, bool createDescriptor) where T : unmanaged
-    {
-        Debug.Assert(count >= 0);
-        return AllocateTempStructuredBuffer<T>((uint)count, createDescriptor);
-    }
-
-    public StructuredBuffer<T> AllocateTempStructuredBuffer<T>(uint count, bool createDescriptor) where T : unmanaged
-    {
-        Debug.Assert(_device != null);
-        //Debug.Assert(int.IsPow2(sizeof(T)), $"A structured buffer must have a size that is a power of 2. Type {typeof(T).Name} has a size of {sizeof(T)}");
-        var stride = (uint)sizeof(T);
-        var size = stride * count;
-        var mappedMemory = GetTempMemory(size, stride);
-        var tempBuffer = new StructuredBuffer<T>
-        {
-            GPUAddress = mappedMemory.GPUAddress,
-            CPUAddress = mappedMemory.CPUAddress,
-            Count = count
-        };
-
-        Debug.Assert(mappedMemory.ResourceOffset % stride == 0);
-
-        if (createDescriptor)
-        {
-            var handle = _heaps[DescriptorHeapType.ShaderResourceView].AllocateTemp();
-            Unsafe.SkipInit(out D3D12_SHADER_RESOURCE_VIEW_DESC srvDesc);
-            srvDesc.Format = DXGI_FORMAT.DXGI_FORMAT_UNKNOWN;
-            srvDesc.ViewDimension = D3D12_SRV_DIMENSION.D3D12_SRV_DIMENSION_BUFFER;
-            srvDesc.Shader4ComponentMapping = D3D12Constants.D3D12_DEFAULT_SHADER_4_COMPONENT_MAPPING;
-            srvDesc.Buffer.FirstElement = mappedMemory.ResourceOffset / stride; // uint32(tempMem.ResourceOffset / stride);
-            srvDesc.Buffer.Flags = D3D12_BUFFER_SRV_FLAGS.D3D12_BUFFER_SRV_FLAG_NONE;
-            srvDesc.Buffer.NumElements = count;
-            srvDesc.Buffer.StructureByteStride = stride;
-            _device.CreateShaderResourceView(_tempBuffers[_frameIndex].Resource, srvDesc, handle.CPU);
-            tempBuffer.DescriptorIndex = handle.Index;
-        }
-
-        return tempBuffer;
-    }
-
-    [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    private MappedTempMemory GetTempMemory(uint size, uint alignment = 0)
-    {
-        var allocSize = size + alignment;
-        var offset = Interlocked.Add(ref _tempBuffers[_frameIndex].Offset, allocSize) - allocSize;
-        Debug.Assert(offset + size < _tempBufferSize);
-        if (alignment > 0)
-        {
-            // align the start offset to the stride of the requested size. 
-            offset = MemoryUtils.AlignNotPowerOf2(offset, alignment);
-        }
-
-        //NOTE(Jens): if we replace the buffers with a single resource (which we should), we need to update the offset to include the frame index
-        return new()
-        {
-            GPUAddress = _tempBuffers[_frameIndex].GPUAddress + offset,
-            CPUAddress = (byte*)_tempBuffers[_frameIndex].CPUAddress + offset,
-            ResourceOffset = offset
-        };
-    }
-
-    public void Update()
-    {
-        _frameIndex = (int)((_frameIndex + 1) % BufferCount);
-        _tempBuffers[_frameIndex].Offset = 0;
-
-        // only the SRV have temporary allocations, no need to call end frame on the other ones.
-        _heaps[DescriptorHeapType.ShaderResourceView].EndFrame();
-    }
-
-    public void Shutdown()
-    {
-        if (_memoryManager != null)
-        {
-            foreach (ref var tempBuffersDescriptor in _tempBuffers)
+            heap.CPUStart = *descriptorHeap->GetCPUDescriptorHandleForHeapStart(MemoryUtils.AsPointer(heap.CPUStart));
+            if (heap.ShaderVisible)
             {
-                tempBuffersDescriptor.Resource.Dispose();
-                tempBuffersDescriptor = default;
+                heap.GPUStart = *descriptorHeap->GetGPUDescriptorHandleForHeapStart(MemoryUtils.AsPointer(heap.GPUStart));
             }
-            for (var i = 0; i < (int)DescriptorHeapType.Count; ++i)
+
+            // Init the free list with indices (we could pre-calculate and store the offsets instead)
+            for (var i = 0; i < heap.NumberOfDescriptors; ++i)
             {
-                _heaps[i].Shutdown(_memoryManager);
+                heap.FreeList[i] = i;
             }
+
+            offset += heap.NumberOfDescriptors;
         }
-        _heaps = default;
-        _memoryManager = null;
     }
 
-    [InlineArray((int)BufferCount)]
-    private struct TempBuffersDescriptors
+
+    public readonly DescriptorHandle Allocate(DescriptorHeapType type)
     {
-        private TempBufferDescriptor _ref;
+        ref var heap = ref *(_heaps.AsPointer() + (int)type);
+        var index = heap.Count++;
+        Debug.Assert(index >= 0);
+        var offset = (uint)(heap.FreeList[index] * heap.IncrementSize);
+
+        var cpuStart = heap.CPUStart.ptr + offset;
+        var gpuStart = heap.ShaderVisible ? heap.GPUStart.ptr + offset : 0ul;
+
+        return new(type, cpuStart, gpuStart, index);
     }
 
-    [SkipLocalsInit]
-    private ref struct MappedTempMemory
+
+    public readonly void Free(in DescriptorHandle handle)
     {
-        public void* CPUAddress;
-        public ulong GPUAddress;
-        public uint ResourceOffset;
+        ref var heap = ref *(_heaps.AsPointer() + (int)handle.Type);
+        var index = --heap.Count;
+
+        heap.FreeList[index] = handle.Index;
+        //TODO(Jens): Add debug check for returning the same handle multiple times.
     }
+
+    [System(SystemStage.Shutdown)]
+    public static void Shutdown(D3D12Allocator* allocator, IMemoryManager memoryManager)
+    {
+        for (var type = 0; type < (int)DescriptorHeapType.Count; ++type)
+        {
+            allocator->_heaps[type].Heap.Dispose();
+        }
+
+        memoryManager.FreeArray(ref allocator->_sharedFreeList);
+        *allocator = default;
+    }
+
+    /// <summary>
+    /// Reset the temporary descriptor heaps at the end of the frame. This is a trivial action so we inline it (runs on main thread by the scheduler)
+    /// </summary>
+    [System(SystemStage.Last, SystemExecutionType.Inline)]
+    public static void Update(ref D3D12Allocator allocator)
+    {
+        allocator._frameIndex = (int)((allocator._frameIndex + 1) % BufferCount);
+
+        //for (var i = 0; i < (int)DescriptorHeapType.Count; ++i)
+        //{
+        //    //TODO(Jens): Implement reset of the temporary buffers (only SRV)
+        //    //allocator->DescriptorHeaps[i].EndFrame();
+        //}
+    }
+
+
 }
