--- src/Titan/Rendering/D3D12/Memory/D3D12AllocatorUpdateSystem.cs
@@ -1,11 +1,0 @@
-
-namespace Titan.Rendering.D3D12.Memory;
-
-internal partial struct D3D12AllocatorUpdateSystem
-{
-    //[Job]
-    //public static void Update(D3D12Allocator allocator)
-    //{
-    //    allocator.Update();
-    //}
-}
