--- src/Titan/Rendering/D3D12New/D3D12Module.cs
@@ -1,25 +1,0 @@
-using Titan.Application;
-using Titan.Rendering.D3D12New.Adapters;
-using Titan.Rendering.D3D12New.Memory;
-
-namespace Titan.Rendering.D3D12New;
-
-internal sealed class D3D12Module : IModule
-{
-    public static bool Build(IAppBuilder builder, AppConfig config)
-    {
-        builder
-#if DEBUG
-            .AddSystemsAndResource<D3D12DebugLayer>()
-#endif
-            .AddSystemsAndResource<D3D12Adapter>()
-            .AddSystemsAndResource<D3D12Device>()
-            .AddSystemsAndResource<D3D12CommandQueue>()
-            .AddSystemsAndResource<D3D12Allocator>()
-
-            .AddSystemsAndResource<DXGISwapchain>()
-            ;
-
-        return true;
-    }
-}
