+++ src/Titan/Systems/SystemDescriptor.cs
@@ -1,4 +1,3 @@
-using System.Collections.Immutable;
 using System.Runtime.InteropServices;
 using Titan.Core.Strings;
